move stubNetworkRequest into utils
