+++ public/js/test/utils/stubNetworkRequest.js
